// src/app/software-seller/order/order-data-interface/verified-order-data-interface.ts
export interface VerifiedOrderDataInterface {
    ButtonView: string;
    Id: number;
    OrderNumber: string;
    OrderDate: string;
    ProductDescription: string;
    ProductPrice: string;
    Quantity: string;
    Amount: string;
    BuyerFullName: string;
    ButtonComplete: string;
    ButtonCancel: string;
  }
  